Fix parsing of node_modules with scoped packages
The `user` group of `urlRegex` would eat beyond the `host` and into the `path`, which isn't valid.
When this happened, it incorrectly assigned everything up to the package scope as the `user`.
Because we weren't expecting a URL-like, we'd strip the `scheme` during parsing, which made it a
path-like, and we don't print the `user` when printing a path-like.

But now we correctly prevent the `user` from extending beyond the `host` of the URL, and the
path-like `node_modules/@babel/runtime/...` will be parsed as part of the `path`.