Fix parsing of file: URLs
File URLs are a trip. They're supposed to be of like: `file:///root/path/file.js` (or
`file:///C:/path/file.js` on Windows). But, the parsing is extremely loose. It doesn't require the
regular `//` in the scheme, nor the leading `/` on the path. And they can parse a host in
`file://foo/bar`, but not in `file://C:/bar`. Just trippy.